implement choosing cursor position just for 1 clause
